put the DatePicker-component-and-friends inside of the TacoView where it needs to go
