+++ client/src/components/LoggedInApp.js
@@ -1,23 +1,11 @@
 
 import React from 'react';
 import Friends from './Friends.js';
-import DatePicker from './TacoLog/DatePicker'
 import TacoView from './TacoLog/TacoView';
 
 // The new App.js! 
 
-
-const style = {
-  position: "relative",
-  margin: "50px auto"
-}
-
 class LoggedInApp extends React.Component {
-
- onDayClick = (e, day) => {
-    alert(day);
-  }
-
   render() {
   return (
     <div className="App">
@@ -25,12 +13,6 @@ class LoggedInApp extends React.Component {
           Let's Get Tacos Homepage
         </p>
       <Friends />
-    
-        <DatePicker style={style} width="302px" 
-          onDayClick={(e, day)=> this.onDayClick(e, day)}/>     
-     
-      {/* <Date /> */}
-
       <TacoView />
     </div>
   );
