--- client/src/components/TacoLog/Date.js
@@ -1,15 +1,0 @@
-import React, { Component } from "react"
-
-class Date extends Component {
-
-
-    render() {
-        return(
-            <h2>This is going to be the date-picker for our taco log!</h2>
-        )
-    }
-}
-
-
-
-export default Date;
